Move Core instances from Conway.Base to Core
