+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance.hs
@@ -3,4 +3,5 @@ module Test.Cardano.Ledger.Conformance (module X) where
 import Test.Cardano.Ledger.Conformance.ExecSpecRule.Core as X
 import Test.Cardano.Ledger.Conformance.Orphans as X ()
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Base as X
+import Test.Cardano.Ledger.Conformance.SpecTranslate.Core as X
 import Test.Cardano.Ledger.Conformance.Utils as X
