+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Base.hs
@@ -27,19 +27,19 @@ module Test.Cardano.Ledger.Conformance.SpecTranslate.Base (
   toSpecRep_,
 ) where
 
-import Cardano.Ledger.Binary (Sized (..))
 import Cardano.Ledger.BaseTypes (Inject (..), NonNegativeInterval, UnitInterval, unboundRational)
+import Cardano.Ledger.Binary (Sized (..))
 import Cardano.Ledger.Compactible (Compactible (..), fromCompact)
-import Data.List.NonEmpty (NonEmpty)
 import Control.DeepSeq (NFData)
 import Control.Monad.Except (ExceptT, MonadError (..), runExceptT)
 import Control.Monad.Reader (MonadReader (..), Reader, asks, runReader)
 import Data.Bitraversable (bimapM)
 import Data.Foldable (Foldable (..))
 import Data.Kind (Type)
-import Data.Maybe.Strict (StrictMaybe (..), strictMaybeToMaybe)
+import Data.List.NonEmpty (NonEmpty)
 import Data.Map (Map)
 import qualified Data.Map as Map
+import Data.Maybe.Strict (StrictMaybe (..), strictMaybeToMaybe)
 import Data.OMap.Strict (OMap)
 import qualified Data.OMap.Strict as OMap
 import Data.OSet.Strict (OSet)
