+++ libs/cardano-ledger-conformance/src/Test/Cardano/Ledger/Conformance/SpecTranslate/Conway.hs
@@ -1,12 +1,6 @@
-module Test.Cardano.Ledger.Conformance.SpecTranslate.Conway (
-  vkeyFromInteger,
-  vkeyToInteger,
-) where
+module Test.Cardano.Ledger.Conformance.SpecTranslate.Conway () where
 
-import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Base (
-  vkeyFromInteger,
-  vkeyToInteger,
- )
+import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Base ()
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Cert ()
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Certs ()
 import Test.Cardano.Ledger.Conformance.SpecTranslate.Conway.Deleg ()
